Created APIs plugin and added user login api.
