+++ plugins/antiques-apis/includes/setting.php
@@ -1,19 +1,9 @@
 <?php
-global $AIT_KEY, $AIT_SECRET, $ALLOWED_ROLES, $SOURCE_LIST, $SOCIAL_MEDIA_TYPES;
+global $API_KEY, $API_SECRET, $ALLOWED_ROLES;
 
-$AIT_KEY = 'DZCxFZdWOjR7';
-$AIT_SECRET = 'Q2P6Tf9cOqYF1aXs8qq7XCtol5TblQT5ZCauAzqUmDAbfh8qbyXHOm5U2pPhmWNyWSrm2NP7';
+$API_KEY = 'DZCxFZdWOjR7ZCxFZd';
+$API_SECRET = '9cOqYF1aXQ2P6Tf9cOqYF1aXs8qq7XCtol5TblQT5ZCauAzqUmDAbfh8qbyXHOm5U2pPhmWNyWSrm2NP7';
 
-$ALLOWED_ROLES = array(	'administrator', 'author', 'editor', 'principleeditor', 'senioreditor');
+$ALLOWED_ROLES = array(	'administrator', 'antique_player', 'antique_vip');
 
-$SOURCE_LIST = array(
-	'backend.bastillepost.com:8282'=>'HONGKONG SANDBOX',
-	'backend.bastillepost.com'=>'HONGKONG',
-	'macau.backend.bastillepost.com'=>'MACAU',
-	'global.backend.bastillepost.com'=>'GLOBAL',
-	'australia.backend.bastillepost.com'=>'AUSTRALIA',
-	'usa.backend.bastillepost.com'=>'USA',	
-);
-
-$SOCIAL_MEDIA_TYPES = array('fan_page','twitter','youtube','fb_ia');
 ?>
